GetIntoTeachingApi/JsonConverters/TrimStringJsonConverter.cs
using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace GetIntoTeachingApi.JsonConverters
{
    public class TrimStringJsonConverter : JsonConverter<string>
    {
        public override bool CanConvert(Type typeToConvert)
        {
            return typeToConvert == typeof(string);
        }

        public override string Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
        {
            string value = reader.GetString();

            return value?.Trim();
        }

        public override void Write(Utf8JsonWriter writer, string value, JsonSerializerOptions options)
        {
            writer.WriteStringValue(value?.Trim());
        }
    }
}